% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sales_ts.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Czyszczenie i przygotowanie szeregów czasowych}
\usage{
clean_sales_ts(
  data,
  date_col = "date",
  sales_col = "sales",
  promo_col = "onpromotion",
  handle_na = "zero",
  aggregate_time = NULL,
  agg_func = sum
)
}
\arguments{
\item{data}{Ramka danych (tibble).}

\item{date_col}{Nazwa kolumny z datą.}

\item{sales_col}{Nazwa kolumny ze sprzedażą.}

\item{promo_col}{Nazwa kolumny z promocjami (domyślnie "onpromotion").}

\item{handle_na}{Sposób obsługi NA: "zero" lub "drop".}

\item{aggregate_time}{Opcjonalna agregacja czasu.}

\item{agg_func}{Funkcja agregacji duplikatów (domyślnie sum).}
}
\value{
Oczyszczona ramka danych.
}
\description{
Sortuje dane, rozwiązuje problem zduplikowanych wpisów agregując je
wybraną funkcją, zachowuje dane o promocjach i obsługuje braki danych (NA).
}
