% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prognosis.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Generowanie prognoz sprzedaży}
\usage{
create_prognosis(data, holidays_data = NULL, horizon = 30, method = "prophet")
}
\arguments{
\item{data}{Oczyszczona ramka danych ze sprzedażą.}

\item{holidays_data}{Ramka danych ze świętami (domyślnie NULL).}

\item{horizon}{Horyzont prognozy w dniach (domyślnie 30).}

\item{method}{Wybór modelu: "arima", "prophet" lub "xgboost".}
}
\value{
Wykres z wygenerowaną prognozą.
}
\description{
Funkcja tworzy prognozy sprzedaży całkowitej przy użyciu modelu ARIMA,
algorytmu Prophet lub uczenia maszynowego XGBoost.
}
