% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sales_data.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Wczytywanie zbiorów danych sprzedażowych}
\usage{
load_sales_data(data_dir)
}
\arguments{
\item{data_dir}{Ścieżka do folderu, w którym znajdują się pobrane pliki CSV.}
}
\value{
Zwraca listę (list) zawierającą trzy obiekty typu \code{tibble}:
\itemize{
\item \code{train} - główne dane sprzedażowe,
\item \code{stores} - metadane sklepów,
\item \code{holidays} - informacje o świętach.
}
}
\description{
Funkcja ładuje z podanego folderu trzy kluczowe pliki:
train.csv, stores.csv oraz holidays_events.csv. Od razu odpowiednio formatuje
ich kolumny
}
