% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sales_trends.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Wizualizacja trendów sprzedażowych}
\usage{
plot_sales_trends(data, date_col = "date", sales_col = "sales")
}
\arguments{
\item{data}{Oczyszczona ramka danych ze sprzedażą.}

\item{date_col}{Nazwa kolumny z datą (domyślnie "date").}

\item{sales_col}{Nazwa kolumny ze sprzedażą (domyślnie "sales").}
}
\value{
Obiekt wykresu typu ggplot.
}
\description{
Generuje wykres liniowy dla szeregu czasowego sprzedaży. Aby wykres był czytelny,
funkcja domyślnie agreguje dane do poziomu tygodniowego. Dodaje również wygładzoną
linię trendu.
}
