% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_ts_logic.R
\name{sales_ts_logic}
\alias{sales_ts_logic}
\title{Główna logika analizy czasowej (Funkcja wyższego rzędu)}
\usage{
sales_ts_logic(
  datasets,
  target_city = NULL,
  target_state = NULL,
  target_type = NULL,
  date_from = NULL,
  date_to = NULL,
  FUN = compute_sales_metrics
)
}
\arguments{
\item{datasets}{Lista z wczytanymi zbiorami danych (wymaga 'train' i 'stores').}

\item{target_city}{Opcjonalnie: nazwa miasta do analizy (np. "Quito").}

\item{target_state}{Opcjonalnie: nazwa stanu do analizy.}

\item{target_type}{Opcjonalnie: typ sklepu (np. "A", "B").}

\item{date_from}{Opcjonalnie: data początkowa w formacie "YYYY-MM-DD".}

\item{date_to}{Opcjonalnie: data końcowa w formacie "YYYY-MM-DD".}

\item{FUN}{Funkcja, która ma zostać wykonana na wyselekcjonowanych danych.
Domyślnie \code{compute_sales_metrics}, ale może to być \code{plot_sales_trends}.}
}
\value{
Wynik działania przekazanej funkcji \code{FUN}.
}
\description{
Funkcja łączy dane sprzedażowe z metadanymi sklepów, filtruje je według
wybranych kryteriów (miasto, stan, typ), a następnie dynamicznie aplikuje
przekazaną funkcję analityczną (np. wyliczanie metryk lub rysowanie wykresu).
}
